Split gatsby-node into separate files
