+++ gatsby/modifyWebpackConfig.js
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ *
+ * @emails react-core
+ */
+
+'use strict';
+
+const {resolve} = require('path');
+const webpack = require('webpack');
+
+module.exports = ({config, stage}) => {
+  // See https://github.com/FormidableLabs/react-live/issues/5
+  config.plugin('ignore', () => new webpack.IgnorePlugin(/^(xor|props)$/));
+
+  config.merge({
+    resolve: {
+      root: resolve(__dirname, '../src'),
+      extensions: ['', '.js', '.jsx', '.json'],
+    },
+  });
+  return config;
+};
